+++ rust_tests/src/bin/update_authorizations.rs
@@ -6,12 +6,14 @@ use ledger::{ApduCommand, LedgerApp};
 
 fn main() {
     let mut key_derivation_path = path::generate_key_derivation_path();
-    let mut update_type = hex::decode("0A").unwrap();
+    let mut update_type = hex::decode("0B").unwrap();
     let mut prefix = hex::decode("02").unwrap();
     let mut public_key_list_length = hex::decode("0002").unwrap();
+    let p2 = 1;
+    let ins = 0x2b;
 
     let ledger = LedgerApp::new().unwrap();
-    
+
     let mut cdata = Vec::new();
     cdata.append(&mut key_derivation_path);
     cdata.append(&mut update_transaction_header::generate_update_transaction_header());
@@ -21,32 +23,39 @@ fn main() {
 
     let command = ApduCommand {
         cla: 224, // Has to be this value for all commands.
-        ins: 0x2A,
+        ins: ins,
         p1: 0,
-        p2: 0,
+        p2: p2,
         length: cdata.len() as u8,
         data: cdata
     };
     ledger.exchange(command).unwrap();
 
     // Send public-keys
-    let public_key_1 = hex::decode(format!("{}{}", "00", "32f892fb3d0dc6138976b6848259cf730e37fa4a61a659c782ec6def978c0828")).unwrap();
-    let public_key_2 = hex::decode(format!("{}{}", "00", "7873cd57848d7aea7be03fbb3f1e8b9e69987fc73f13e473356776a16f26c96b")).unwrap();
+    let public_key_1 = hex::decode(format!("{}{}", "00", "b6bc751f1abfb6440ff5cce27d7cdd1e7b0b8ec174f54de426890635b27e7daf")).unwrap();
+    let public_key_2 = hex::decode(format!("{}{}", "00", "46a3e38ddf8b493be6e979034510b91db5448da9cba48c106139c288d658a004")).unwrap();
     let public_key_list = [public_key_1, public_key_2];
 
     for public_key in public_key_list.iter() {
         let command = ApduCommand {
             cla: 224,
-            ins: 0x2A,
+            ins: ins,
             p1: 1,
-            p2: 0,
+            p2: p2,
             length: 0,
             data: public_key.clone()
         };
         ledger.exchange(command).unwrap();
     }
+    let mut result = None;
 
-    for i in 0..12 {
+    let structure_count = if p2 == 0 {
+        12
+    } else {
+        14
+    };
+
+    for _ in 0..structure_count {
         // Access structure
         let mut size = hex::decode("0003").unwrap();
 
@@ -55,9 +64,9 @@ fn main() {
 
         let command = ApduCommand {
             cla: 224, // Has to be this value for all commands.
-            ins: 0x2A,   // Sign update authorizations
+            ins: ins,   // Sign update authorizations
             p1: 2,
-            p2: 0,
+            p2: p2,
             length: command_data.len() as u8,
             data: command_data
         };
@@ -75,9 +84,9 @@ fn main() {
 
         let command = ApduCommand {
             cla: 224,
-            ins: 0x2A,
+            ins: ins,
             p1: 3,
-            p2: 0,
+            p2: p2,
             length: access_structure_cdata.len() as u8,
             data: access_structure_cdata
         };
@@ -89,15 +98,15 @@ fn main() {
 
         let command = ApduCommand {
             cla: 224,
-            ins: 0x2A,
+            ins: ins,
             p1: 4,
-            p2: 0,
+            p2: p2,
             length: threshold_cdata.len() as u8,
             data: threshold_cdata
         };
-        let result = ledger.exchange(command).unwrap();
-        if i == 11 {
-            println!("Signature: {}", hex::encode(result.data));
-        }
+        result = Some(ledger.exchange(command).unwrap());
+    }
+    if let Some(r) = result {
+        println!("Signature: {}", hex::encode(r.data));
     }
 }