+++ tests/src/updateAuthorizations.test.ts
@@ -5,6 +5,7 @@ import { setupZemu } from './options';
 async function updateAuthorizations(
     sim: Zemu, transport: Transport,
     ins: number,
+    p2: number,
     type: string,
     prefix: string,
     expectedSignature: string,
@@ -17,7 +18,7 @@ async function updateAuthorizations(
         Buffer.from('0002', 'hex'),
     ]);
 
-    transport.send(0xe0, ins, 0x00, 0x00, data);
+    transport.send(0xe0, ins, 0x00, p2, data);
     await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
     await sim.clickRight();
     let snapshot = await sim.clickBoth();
@@ -27,22 +28,23 @@ async function updateAuthorizations(
     const keys = [key1, key2];
 
     for (const key of keys) {
-        transport.send(0xe0, ins, 0x01, 0x00, key);
+        transport.send(0xe0, ins, 0x01, p2, key);
         await sim.waitUntilScreenIsNot(snapshot);
         snapshot = await handleKeyUi();
     }
 
+    const structureCount = p2 ? 14 : 12;
     // Go through each access structure
-    for (let i = 0; i < 12; i += 1) {
+    for (let i = 0; i < structureCount; i += 1) {
         const accessStructureSize = Buffer.from('0003', 'hex');
-        await transport.send(0xe0, ins, 0x02, 0x00, accessStructureSize);
+        await transport.send(0xe0, ins, 0x02, p2, accessStructureSize);
 
         const keyIndex1 = Buffer.from('0001', 'hex');
         const keyIndex2 = Buffer.from('00F1', 'hex');
         const keyIndex3 = Buffer.from('000C', 'hex');
 
         const accessStructureData = Buffer.concat([keyIndex1, keyIndex2, keyIndex3]);
-        transport.send(0xe0, ins, 0x03, 0x00, accessStructureData);
+        transport.send(0xe0, ins, 0x03, p2, accessStructureData);
         await sim.waitUntilScreenIsNot(snapshot);
         await sim.clickBoth();
         await sim.clickBoth();
@@ -50,11 +52,10 @@ async function updateAuthorizations(
         snapshot = await sim.clickBoth();
 
         const threshold = Buffer.from('0002', 'hex');
-        const tx = transport.send(0xe0, ins, 0x04, 0x00, threshold);
+        const tx = transport.send(0xe0, ins, 0x04, p2, threshold);
         await sim.waitUntilScreenIsNot(snapshot);
         snapshot = await sim.clickBoth();
-
-        if (i === 11) {
+        if (i === structureCount - 1) {
             await sim.clickBoth();
             await expect(tx).resolves.toEqual(
                 Buffer.from(expectedSignature, 'hex'),
@@ -63,11 +64,12 @@ async function updateAuthorizations(
     }
 }
 
-test('[NANO S] Update level 2 keys with root keys', setupZemu('nanos', async (sim, transport) => {
+test('[NANO S] Update level 2 keys with root keys v0', setupZemu('nanos', async (sim, transport) => {
     await updateAuthorizations(
         sim,
         transport,
         0x2a,
+        0x00,
         '0a',
         '02',
         'e9740c83110c2bc4fe876da81050cff27c1ab282178a3049327c6e97ebea34f2cd5f038e2b8dff466b1455e4d667d95d050fed1af04ea80c40360854dede12049000',
@@ -81,11 +83,12 @@ test('[NANO S] Update level 2 keys with root keys', setupZemu('nanos', async (si
     );
 }));
 
-test('[NANO X] Update level 2 keys with root keys', setupZemu('nanox', async (sim, transport) => {
+test('[NANO X] Update level 2 keys with root keys v0', setupZemu('nanox', async (sim, transport) => {
     await updateAuthorizations(
         sim,
         transport,
         0x2a,
+        0x00,
         '0a',
         '02',
         'e9740c83110c2bc4fe876da81050cff27c1ab282178a3049327c6e97ebea34f2cd5f038e2b8dff466b1455e4d667d95d050fed1af04ea80c40360854dede12049000',
@@ -96,11 +99,12 @@ test('[NANO X] Update level 2 keys with root keys', setupZemu('nanox', async (si
     );
 }));
 
-test('[NANO S] Update level 2 keys with level 1 keys', setupZemu('nanos', async (sim, transport) => {
+test('[NANO S] Update level 2 keys with level 1 keys v0', setupZemu('nanos', async (sim, transport) => {
     await updateAuthorizations(
         sim,
         transport,
         0x2b,
+        0x00,
         '0b',
         '01',
         'b3843361c16bc8f9f7792df766f7192e8b77aae420c958d290d8408b9ac224350998d4c264e4e60c577385963a574391b122948afae0dcb1e0ae7fda8076150e9000',
@@ -114,14 +118,85 @@ test('[NANO S] Update level 2 keys with level 1 keys', setupZemu('nanos', async 
     );
 }));
 
-test('[NANO X] Update level 2 keys with level 1 keys', setupZemu('nanox', async (sim, transport) => {
+test('[NANO X] Update level 2 keys with level 1 keys v0', setupZemu('nanox', async (sim, transport) => {
     await updateAuthorizations(
         sim,
         transport,
         0x2b,
+        0x00,
         '0b',
         '01',
         'b3843361c16bc8f9f7792df766f7192e8b77aae420c958d290d8408b9ac224350998d4c264e4e60c577385963a574391b122948afae0dcb1e0ae7fda8076150e9000',
+        async () => {
+            await sim.clickRight();
+            return sim.clickBoth();
+        },
+    );
+}));
+
+test('[NANO S] Update level 2 keys with root keys v1', setupZemu('nanos', async (sim, transport) => {
+    await updateAuthorizations(
+        sim,
+        transport,
+        0x2a,
+        0x01,
+        '0a',
+        '03',
+        '02f76c94c76d552364e8b054001f86c2cd4417de90343920558ac00529e2bc429f6db1a37c7f75fea69ec663bfadcc847ad568996545de93adfa1674b12d41079000',
+        async () => {
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            return sim.clickBoth();
+        },
+    );
+}));
+
+test('[NANO X] Update level 2 keys with root keys v1', setupZemu('nanox', async (sim, transport) => {
+    await updateAuthorizations(
+        sim,
+        transport,
+        0x2a,
+        0x01,
+        '0a',
+        '03',
+        '02f76c94c76d552364e8b054001f86c2cd4417de90343920558ac00529e2bc429f6db1a37c7f75fea69ec663bfadcc847ad568996545de93adfa1674b12d41079000',
+        async () => {
+            await sim.clickRight();
+            return sim.clickBoth();
+        },
+    );
+}));
+
+test('[NANO S] Update level 2 keys with level 1 keys v1', setupZemu('nanos', async (sim, transport) => {
+    await updateAuthorizations(
+        sim,
+        transport,
+        0x2b,
+        0x01,
+        '0b',
+        '02',
+        'ab92d78f158730042a1a4c2f738600f064ad5100551afad3760053f8f9e28e12d47a4e0484417f879e065332612e3cc4c5315e4f610f3fd4e3d7469c058556019000',
+        async () => {
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            return sim.clickBoth();
+        },
+    );
+}));
+
+test('[NANO X] Update level 2 keys with level 1 keys v1', setupZemu('nanox', async (sim, transport) => {
+    await updateAuthorizations(
+        sim,
+        transport,
+        0x2b,
+        0x01,
+        '0b',
+        '02',
+        'ab92d78f158730042a1a4c2f738600f064ad5100551afad3760053f8f9e28e12d47a4e0484417f879e065332612e3cc4c5315e4f610f3fd4e3d7469c058556019000',
         async () => {
             await sim.clickRight();
             return sim.clickBoth();